23/11

+++ src/main/java/edu/hcmuaf/fit/nlpige/recognize/multiplechoicesheet/tool/supporter/form/FormUtils.java
@@ -5,15 +5,19 @@ import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.PageSize;
 import com.itextpdf.text.pdf.PdfWriter;
 import com.itextpdf.tool.xml.XMLWorkerHelper;
+import edu.hcmuaf.fit.nlpige.recognize.multiplechoicesheet.common.logging.Logger;
 
 import java.io.*;
 import java.nio.charset.Charset;
+import java.util.Calendar;
+import java.util.Date;
 
-public class FormUtils {
+public class FormUtils extends Logger {
 
     public static final String HEAD_PATH = "src/main/resources/default/html-head.txt";
     public static final String HEADER_PATH = "src/main/resources/default/html-header.txt";
     public static final String CONTENT_PATH = "src/main/resources/default/html-content.txt";
+    public static final String FORM_NAME = "form.html";
 
     public static FormHeader getFormHeader(String file) {
         try {
@@ -27,8 +31,10 @@ public class FormUtils {
             return formHeader;
         } catch (FileNotFoundException e) {
             e.printStackTrace();
+            log.error("File header not found!", e);
         } catch (IOException e) {
             e.printStackTrace();
+            log.error("Reading file error!", e);
         }
         return null;
     }
@@ -51,10 +57,14 @@ public class FormUtils {
         return null;
     }
 
-    public static void generateHtml(FormHeader formHeader, FormBody formBody, String output) {
+    public static void generateHtml(FormHeader formHeader, FormBody formBody, String desFolder) {
         try {
             BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(HEAD_PATH)));
-            PrintWriter pw = new PrintWriter(new FileOutputStream(output));
+
+            String formPath = desFolder  + "/" + FORM_NAME;
+
+            FileOutputStream fos = new FileOutputStream(formPath);
+            PrintWriter pw = new PrintWriter(fos);
             String line;
             while ((line = br.readLine()) != null) {
                 pw.write(line);
@@ -109,7 +119,7 @@ public class FormUtils {
         }
     }
 
-    public static void setQRCode(String srcFile, String qrCode) {
+    public static void setQRCode(String srcFile,String desFolder, String qrCode) {
         try {
             BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(srcFile)));
             StringBuilder document = new StringBuilder();
@@ -123,7 +133,9 @@ public class FormUtils {
             int index = document.indexOf("%s");
             document.replace(index, index+2, qrCode);
 
-            PrintWriter pw = new PrintWriter(new FileOutputStream(srcFile));
+            String dest = desFolder + "/" + System.currentTimeMillis() + ".html";
+
+            PrintWriter pw = new PrintWriter(new FileOutputStream(dest));
             pw.write(document.toString());
             pw.flush();
             pw.close();
@@ -134,35 +146,41 @@ public class FormUtils {
         }
     }
 
-    public static void convertXhtmlToPdf(String srcFile, String destFolder) {
+    public static void convertXhtmlToPdf(String srcFolder, String desFolder) {
         try {
-            File src = new File(srcFile);
+            File src = new File(srcFolder);
             if (!src.exists()) {
-
+                throw new FileNotFoundException("File not exists!");
             }
-            // create pdf document
-            Document document = new Document(PageSize.A4);
-            // open file input stream
-            FileInputStream fis = new FileInputStream(src);
 
-            // convert name to .pdf
-            StringBuffer fileName = new StringBuffer(src.getName());
-            fileName.delete(fileName.lastIndexOf("."), fileName.length());
-            fileName.append(".pdf");
+            File[] subFiles = src.listFiles();
 
-            // open file output stream
-            FileOutputStream fos = null;
+            if(subFiles == null) {
+                throw new FileNotFoundException("Folder is empty!");
+            }
 
-            fos = new FileOutputStream(new File(destFolder, fileName.toString()));
+            for (File file: subFiles) {
+                // create pdf document
+                Document document = new Document(PageSize.A4);
+                // open file input stream
+                FileInputStream fis = new FileInputStream(file);
 
+                // convert name to .pdf
+                StringBuffer fileName = new StringBuffer(file.getName());
+                fileName.delete(fileName.lastIndexOf("."), fileName.length());
+                fileName.append(".pdf");
 
-            // create writer
-            PdfWriter writer = PdfWriter.getInstance(document, fos);
+                // open file output stream
+                FileOutputStream fos = new FileOutputStream(new File(desFolder, fileName.toString()));
 
-            // convert xhtml to pdf
-            document.open();
-            XMLWorkerHelper.getInstance().parseXHtml(writer, document, fis, Charset.forName("UTF-8"));
-            document.close();
+                // create writer
+                PdfWriter writer = PdfWriter.getInstance(document, fos);
+
+                // convert xhtml to pdf
+                document.open();
+                XMLWorkerHelper.getInstance().parseXHtml(writer, document, fis, Charset.forName("UTF-8"));
+                document.close();
+            }
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         } catch (DocumentException e) {
